GRAY_CC is short
[[Imported from SVN: r5556]]

+++ dev/ps/postscript.c
@@ -69,7 +69,7 @@
 
 #define COLORS          256
 #define GRAY                    0.5
-#define GRAY_CC                 (~(unsigned long)0)
+#define GRAY_CC                 (~(short)0)
 
 #define TRFMX(pt) (((float)(pt.x))*mxx + ((float)(pt.y))*mxy + tx)
 #define TRFMY(pt) (((float)(pt.x))*myx + ((float)(pt.y))*myy + ty)
